HexEdit: * Made the "Value Preview" option enabled by default * Some small project adjustments * Removed some warnings
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@5140 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/HexEditor/FileContentBuffered.cpp
@@ -268,7 +268,7 @@ bool FileContentBuffered::ReadFile(const wxString& fileName)
     m_UndoCurrent = 0;
     m_UndoSaved = 0;
 
-    return fl.Read( &m_Buffer[0], m_Buffer.size() ) == m_Buffer.size();
+    return (size_t)fl.Read( &m_Buffer[0], m_Buffer.size() ) == m_Buffer.size();
 }
 
 const FileContentBuffered::ExtraUndoData* FileContentBuffered::Redo()
